+++ app/Http/Controllers/GalleryController.php
@@ -4,8 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\GalleryRequest;
 use App\Models\Event;
-use Illuminate\Http\Request;
 use App\Models\Gallery;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class GalleryController extends Controller
@@ -28,18 +28,18 @@ class GalleryController extends Controller
     public function create()
     {
         $evenementen = Event::orderBy('datum', 'desc')->get();
+
         return view('gallery.create', compact('evenementen'));
     }
 
     public function store(GalleryRequest $request)
     {
+        $validated = $request->validated();
+
         foreach ($request->file('images') as $image) {
             $imagePath = $image->store('gallery', 'public');
-
             $photo = Gallery::create([
-                'title' => $request->validated(['title']),
-                'date' => $request->validated(['date']),
-                'type' => $request->validated(['type']),
+                ...$validated,
                 'src' => $imagePath,
             ]);
 
@@ -48,32 +48,23 @@ class GalleryController extends Controller
             }
         }
 
-        return redirect()->route('gallery.index')->with('success', 'Foto\'s succesvol toegevoegd');
+        return redirect()
+            ->route('gallery.index')
+            ->with('success', 'Foto\'s succesvol toegevoegd');
     }
 
     public function edit(Gallery $gallery)
     {
         $evenementen = Event::orderBy('datum', 'desc')->get();
+
         return view('gallery.edit', compact('gallery', 'evenementen'));
     }
 
     public function update(GalleryRequest $request, Gallery $gallery)
     {
-        $gallery->title = $request->validated(['title']);
-        $gallery->date = $request->validated(['date']);
-        $gallery->type = $request->validated(['type']);
-
-        if ($request->hasFile('image')) {
-            if ($gallery->src) {
-                Storage::delete($gallery->src);
-            }
-
-            $imagePath = $request->file('image')->store('gallery', 'public');
-
-            $gallery->src = $imagePath;
-        }
-
-        $gallery->save();
+        $validated = $request->validated();
+        $request->hasFile('image') && $validated['image'] = $gallery->replaceFile($request->file('image'), 'gallery', 'public', 'image');
+        $gallery->update($validated);
         $gallery->evenementen()->sync($request->input('evenementen', []));
 
         return redirect()
@@ -83,10 +74,7 @@ class GalleryController extends Controller
 
     public function destroy(Gallery $photo)
     {
-        if ($photo->src) {
-            Storage::delete($photo->src);
-        }
-
+        $photo->src && Storage::delete($photo->src);
         $photo->delete();
 
         return redirect()
